Remove `egg-mode` and `tweetust` features
Those features make `egg-mode` and `tweetust` crates public
dependencies, meaning that we need a minor version bump in order to
update the dependencies, which is a compatibility hazard.

Additionally, this commit removes `tweetust` dependency from the example
code. The crate no longer seems to be maintained actively and is
unlikely unlikely to update its HTTP stack (`hyper`) at least in the
foreseeable future. Using `tweetust` feature undesirably causes the
dependency graph to have two version of `hyper` simultaneously, making
the crate an unsuitable example (the same goes for `egg-mode`, but that
crate is more likely to migrate to `futures` 0.3 ecosystem in the
not-too-distant future, so the problem should be transient).