Remove documentation on MSRV
The crate has a bunch of (transitive) dependencies and it is quite
difficult to be certain that all of those dependencies keep their MSRV
between "SemVer-compatible" versions.

We'll keep a test for a minimum Rust version on CI though. However, it is
only there to discourage us from unconsciously making excessive bumps of
the minimum compatible Rust version, not to prevent all possible sorts
of version bumps.