Use `memmem` to simplify the code
This might also improve the performance, but I don't know. The
implimentation of `memmem` looks quite sophisticated so it's difficult
to guess the perf impact w/o running a benchmark (that doesn't mean it's
slow of course!). Anyway, that doesn't matter so much since the point of
this change is to simplify our code.